+++ 02_fs文件系統/代碼/11-兩種路徑.js
@@ -1,0 +1,10 @@
+// 1. 導入 fs 模塊
+const fs = require("fs");
+
+// 相對路徑
+// fs.writeFileSync("./index.html", "love");
+// fs.writeFileSync("index.html", "love");
+// fs.writeFileSync("../index.html", "love");
+
+// 絕對路徑
+// fs.writeFileSync("/Users/xuyuanshi/Desktop/index.html", "love");
